refactor middleware
