+++ apps/sps-ordbokene-web/components/custom-pages/word-view.vue
@@ -97,7 +97,7 @@ useHead({
 
 definePageMeta({
     middleware: [
-      "simple-search"
+      "simple-search-middleware"
     ]
   })
