+++ apps/sps-ordbokene-web/middleware/simple-search-middleware.ts
@@ -1,0 +1,25 @@
+import { useSearchStore } from '~/stores/searchStore'
+export default defineNuxtRouteMiddleware(async (to, from) => {
+    const store = useSearchStore()
+    // Redirect if javascript is disabled
+    if (to.query.q) {
+      return `/${to.params.locale}/${to.params.dict}/${to.query.q}`
+    }
+
+    store.dict = to.params.dict || 'bm,nn'
+    store.q = to.params.q || to.query.q
+    if (store.q) {
+      store.searchUrl = to.fullPath
+      store.lemmas.bm = new Set()
+      store.lemmas.nn = new Set()
+
+      // Redirect to advanced search if special symbols
+      if (advancedSpecialSymbols(store.q)) {
+        return navigateTo(`/${to.params.locale}/search?q=${store.q}&dict=${to.params.dict}&scope=${store.scope || 'ei'}`)
+      }
+
+      store.input = to.query.orig || store.q
+      
+
+    }
+})
