+++ apps/sps-ordbokene-web/middleware/welcome-middleware.ts
@@ -1,0 +1,5 @@
+import { useSearchStore } from '~/stores/searchStore'
+export default defineNuxtRouteMiddleware(async (to, from) => {
+    const store = useSearchStore()
+    store.$reset()
+})
